pyproject.toml
[tool.poetry]
name = "fastapi-load-test"
version = "0.1.0"
description = ""
authors = ["Jingwei Kang"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.112.1"}
aiohttp = "^3.10.5"
requests = "^2.32.3"
opentelemetry-instrumentation-fastapi = "^0.47b0"
opentelemetry-api = "^1.26.0"
opentelemetry-sdk = "^1.26.0"
opentelemetry-exporter-otlp-proto-common = "^1.26.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.26.0"
